refactor(cli): enable --file for mdk gen +
