--- meta-cli/src/cli/generators/mod.rs
@@ -1,5 +1,0 @@
-// Copyright Metatype OÜ, licensed under the Mozilla Public License Version 2.0.
-// SPDX-License-Identifier: MPL-2.0
-
-pub mod codegen;
-pub mod metagen;
